transition: removed most of the existing code, to have lean starter, also fixed a weird behavior with MSVC and CMake
