+++ src/Version.h.in
@@ -1,2 +1,3 @@
-#define ReseauInterurbain_VERSION_MAJOR @ReseauInterurbain_VERSION_MAJOR@
-#define ReseauInterurbain_VERSION_MINOR @ReseauInterurbain_VERSION_MINOR@
+#define APPLICATION_VERSION_MAJOR ${CMAKE_PROJECT_VERSION_MAJOR}
+#define APPLICATION_VERSION_MINOR ${CMAKE_PROJECT_VERSION_MINOR}
+#define APPLICATION_NAME L"${PROJECT_NAME}"
